feat: use authorized user id to resolve his related data

+++ src/app/controllers/v1/profiles/profiles.controller.ts
@@ -18,7 +18,7 @@ import * as dayjs from 'dayjs';
 
 import { V1_API_TAGS } from '@/app/constants';
 import { ApiVersion, ControllerVersionEnum, parseUuid } from '@/common';
-import { DJwtPayload, JwtPayload } from '@/modules/auth';
+import { AuthorizedId } from '@/modules/auth';
 import {
   BirthdayMarkerEnum,
   ProfilesOrderEnum,
@@ -53,11 +53,12 @@ export class ProfilesControllerV1 {
     type: [response.ProfileDto],
   })
   async getMany(
-    @DJwtPayload() jwtPayload: JwtPayload,
+    @AuthorizedId() authorizedId: string,
   ): Promise<response.ProfileDto[]> {
-    const userTz = await this.usersService.getTimeZone(jwtPayload.id);
+    const userTz = await this.usersService.getTimeZone(authorizedId);
 
     const profiles = await this.profilesService.getMany({
+      userId: authorizedId,
       order: ProfilesOrderEnum.Upcoming,
       tz: userTz ?? undefined,
     });
@@ -86,14 +87,22 @@ export class ProfilesControllerV1 {
   @ApiOkResponse({
     type: response.ProfileDto,
   })
-  async getById(@Param('id') id: string): Promise<response.ProfileDto> {
+  async getById(
+    @AuthorizedId() authorizedId: string,
+    @Param('id') id: string,
+  ): Promise<response.ProfileDto> {
     const uuid = parseUuid(id);
-    const profile = await this.profilesService.get(uuid);
 
-    if (!profile) {
+    const isBelongToUser = await this.profilesService.isBelongToUser(
+      id,
+      authorizedId,
+    );
+
+    if (!isBelongToUser) {
       throw new NotFoundException();
     }
 
+    const profile = await this.profilesService.get(uuid);
     const isYear = profile.birthdayMarker === BirthdayMarkerEnum.Standard;
 
     return {
@@ -111,10 +120,13 @@ export class ProfilesControllerV1 {
   @ApiBody({
     type: request.CreateDto,
   })
-  async createOne(@Body() dto: request.CreateDto): Promise<void> {
+  async createOne(
+    @AuthorizedId() authorizedId: string,
+    @Body() dto: request.CreateDto,
+  ): Promise<void> {
     const { name, birthday } = dto;
 
-    await this.profilesService.create({
+    await this.profilesService.create(authorizedId, {
       name,
       birthday,
     });
@@ -128,8 +140,20 @@ export class ProfilesControllerV1 {
     name: 'id',
     type: 'uuid',
   })
-  async deleteById(@Param('id') id: string): Promise<void> {
+  async deleteById(
+    @AuthorizedId() authorizedId: string,
+    @Param('id') id: string,
+  ): Promise<void> {
     const uuid = parseUuid(id);
+
+    const isBelongToUser = await this.profilesService.isBelongToUser(
+      id,
+      authorizedId,
+    );
+
+    if (!isBelongToUser) {
+      throw new NotFoundException();
+    }
 
     await this.profilesService.delete(uuid);
   }
